Trigger clickOutside first!
I'm having this funny issue with states when playing with elements that are shown/hidden by an "outsider" element. 

```html
<div id="sidebar" data-controller="sidebar"></div>
<div id="outside_burger_that_opens_the_sidebar"></div>
```

The `burger` element opens the sidebar, being outside the `sidebar` controller/element scope. The `sidebar` controller uses `clickOutside` to close/hide the sidebar if it's open. When clicking the burger, its changes the state of the sidebar and opens it making it visible with in the viewport. And then `clickOutside` is triggered closing it again (as it's visible and the `state` just changed). `open -> clickOutside -> close`

As workaround I've been adding a delay of 1ms to execute the lines that changes the state of the sidebar, then `clickOutside` don't even trigger as the sidebar it's not visible or the state hasn't changed neither.

But I think `capture` the click event and trigger `clickOutside` first is the way to go, as this will make this events trigger before any changes on the controller element state or visibility. And I don't see any case then capturing the event could cause issues.